Move StoreWriter#writeToStore into test helpers module.
