+++ src/cache/inmemory/__tests__/roundtrip.ts
@@ -6,6 +6,7 @@ import { EntityStore } from '../entityStore';
 import { StoreReader } from '../readFromStore';
 import { StoreWriter } from '../writeToStore';
 import { InMemoryCache } from '../inMemoryCache';
+import { writeQueryToStore } from './helpers';
 
 function assertDeeplyFrozen(value: any, stack: any[] = []) {
   if (value !== null && typeof value === 'object' && stack.indexOf(value) < 0) {
@@ -29,7 +30,8 @@ function storeRoundtrip(query: DocumentNode, result: any, variables = {}) {
   const reader = new StoreReader({ cache });
   const writer = new StoreWriter(cache);
 
-  const store = writer.writeQueryToStore({
+  const store = writeQueryToStore({
+    writer,
     result,
     query,
     variables,
@@ -67,7 +69,8 @@ function storeRoundtrip(query: DocumentNode, result: any, variables = {}) {
 
   // Now make sure subtrees of the result are identical even after we write
   // an additional bogus field to the store.
-  writer.writeQueryToStore({
+  writeQueryToStore({
+    writer,
     store,
     result: { oyez: 1234 },
     query: gql`
